fix: fixes multiple issues

+++ src/mesh.hpp
@@ -20,14 +20,14 @@ public:
     void LoadMesh(const char *path); // Load mesh from file
     void Draw();                     // Draw the mesh
 
-private:                                 // Private Methods
-    std::vector<Vertex> Interlace();     // Interlace vertex data for GPU
-    void SetupMesh();                    // Setup the mesh buffers
-private:                                 // Private members
-    std::vector<glm::vec3> m_positions;  // Vertex data
-    std::vector<glm::vec3> m_normals;    // Normal data
-    std::vector<glm::vec2> m_uvs;        // Texture coordinate data
-    std::vector<unsigned int> m_indices; // Index data
+private:                                // Private Methods
+    std::vector<Vertex> Interlace();    // Interlace vertex data for GPU
+    void SetupMesh();                   // Setup the mesh buffers
+private:                                // Private members
+    std::vector<glm::vec3> m_positions; // Vertex data
+    std::vector<glm::vec3> m_normals;   // Normal data
+    std::vector<glm::vec2> m_uvs;       // Texture coordinate data
+    std::vector<uint16_t> m_indices;    // Index data
 
     unsigned int m_VAO; // Vertex Array Object ID
     unsigned int m_VBO; // Vertex Buffer Object ID
